requirements.txt
pygame-ce
pyyaml
cerberus
